+++ wp-content/plugins/cat-game-vote-standalone/templates/pages/leaderboard.php
@@ -65,6 +65,7 @@ $current_user_id = (int) ($data['current_user_id'] ?? 0);
                     <p class="cg-location">📍 <?php echo esc_html($item['city'] . ', ' . $item['country']); ?></p>
                     <div class="cg-rank-reactions">
                         <?php CatGame_Reactions::render_widget((int) ($item['id'] ?? 0), is_user_logged_in(), (array) ($item['reaction_counts'] ?? []) ? ['reaction_counts' => (array) ($item['reaction_counts'] ?? []), 'my_reaction' => ($item['my_reaction'] ?? null)] : []); ?>
+                        <?php echo class_exists('CatGame_Reports') ? CatGame_Reports::report_button_html($item, $current_user_id) : ''; ?>
                     </div>
                 </div>
             </article>